LunaPhase3/pyproject.toml
[tool.poetry]
name = "lunaphase3"
version = "0.1.0"
description = ""
authors = ["Kristian Mischke <kristianmischke@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.4"
python-dotenv = "^1.0.0"
discord-py = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
